Update gainphaseoffset.m
Updated function. The question wether the entire trace should be used has been removed, because there might be always some filtering artifacts that could compromise behavioural analysis. 
It allows to select that part of the EODf that looks most stationary. Keep in mind that we need at least 4 cycles of response (for envelope frequencies > 1Hz, it's at least 10 cycles, because the last ten will be removed during processing).