Update distFunc.m
Myriah found a minor bug in that code. Here is what she wrote to me:

"The issue is that when populating the combCh which is used as an index matrix, if there are neurons on Ch0 of the neuropixel this cannot be used as an index. The easiest fix I could find for this was simply to set any zeros to 1... It wont be as precise of a "location" relative to the NPIX probe but it's as close as possible... So I added the following code to line 10:
combCh(combCh == 0) = 1;

Obviously we don't want to shift all the channels by 1, because their relative position would be inaccurate... so I figured it was better to have one channel assumed to be in the wrong place rather than all of them."